requirements.txt
pyTelegramBotAPI==4.14.0
Flask==3.0.0
pymongo==4.6.2
dnspython==2.6.1
qrcode==7.4.2
schedule==1.2.1
gunicorn==21.2.0
python-dotenv==1.0.0
Pillow==10.3.0
requests==2.31.0
